fix(api): land SPEC-087 stats N+1 and shared guest identity (#334 #335)
Replace workspace-stats embedding N+1 with KVStorage aggregate + relational
chunks SSOT, and stop minting per-browser anon_* users in favor of a shared
per-tenant guest with JWT user_id bind and admin filtering.

+++ edgequake/crates/edgequake-api/src/handlers/documents/recovery/stuck.rs
@@ -66,9 +66,8 @@ pub(crate) async fn run_recover_stuck(
     );
 
     // SPEC-086: fail orphan staging shells (list-visible) before final-metadata recover.
-    let staging_age = std::time::Duration::from_secs(
-        (request.stuck_threshold_minutes as u64).saturating_mul(60),
-    );
+    let staging_age =
+        std::time::Duration::from_secs(request.stuck_threshold_minutes.saturating_mul(60));
     if let Err(e) = crate::services::recover_orphaned_staging_admissions(
         std::sync::Arc::clone(&state.storage.kv_storage),
         std::sync::Arc::clone(&state.tasks.storage),
@@ -91,11 +90,12 @@ pub(crate) async fn run_recover_stuck(
 
     // P-G7 + SPEC-027: batch scoped metadata (suffix index + tenant filter).
     // Prefer progress loader so aged staging shells (post 086 merge) are visible.
-    let scoped_metadata = crate::services::document_metadata_scan::load_scoped_document_metadata_for_progress(
-        state.storage.kv_storage.as_ref(),
-        &tenant_ctx,
-    )
-    .await?;
+    let scoped_metadata =
+        crate::services::document_metadata_scan::load_scoped_document_metadata_for_progress(
+            state.storage.kv_storage.as_ref(),
+            &tenant_ctx,
+        )
+        .await?;
 
     let mut stuck_docs = Vec::new();
     let mut requeued_ids = Vec::new();
